QUBOConstraints

Documentation for QUBOConstraints.

QUBOConstraints.AbstractOptimizer

QUBOConstraints.QUBO_base

QUBOConstraints.QUBO_linear_sum

QUBOConstraints.binarize

QUBOConstraints.debinarize

QUBOConstraints.is_valid

QUBOConstraints.train

QUBOConstraints.AbstractOptimizer — Type

AbstractOptimizer

An abstract type (interface) used to learn QUBO matrices from constraints. Only a train method is required.

QUBOConstraints.QUBO_base — Function

QUBO_base(n, weight = 1)

A basic QUBO matrix to ensure that binarized variables keep a valid encoding.

QUBOConstraints.QUBO_linear_sum — Method

QUBO_linear_sum(n, σ)

One valid QUBO matrix given n variables and parameter σ for the linear sum constraint.

QUBOConstraints.binarize — Method

binarize(x[, domain]; binarization = :one_hot)

Binarize x following the binarization encoding. If x is a vector (instead of a number per say), domain is optional.

QUBOConstraints.debinarize — Method

debinarize(x[, domain]; binarization = :one_hot)

Transform a binary vector into a number or a set of number. If domain is not given, it will compute a default value based on binarization and x.

QUBOConstraints.is_valid — Function

is_valid(x, encoding::Symbol = :none)

Check if x has a valid format for encoding.

For instance, if encoding == :one_hot, at most one bit of x can be set to 1.

QUBOConstraints.train — Method

train(args...)

Default train method for any AbstractOptimizer.